requirements.txt
torch>=2.0.0
transformers>=4.38.0
datasets>=2.18.0
tokenizers>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: experiment tracking
# wandb>=0.16.0

# Optional: faster tokenizers
# sentencepiece>=0.1.99

# Optional: Kaggle dataset download
# kaggle>=1.6.0
